- Update live room for API remove user  - Can remove user when IsRoom = true

+++ catalog/model/friend/room.php
@@ -58,7 +58,7 @@ class ModelFriendRoom extends Model {
 		}
 
 		// user_id is remove user
-		if ( !empty($aData['user_id']) ) {
+		if ( !empty($aData['user_id']) && $oRoom->getIsRoom() ) {
 			if ( $oRoom->getCreator()->getId() == $aData['user_id'] || $oRoom->getUsers()->count() == 2 ) {
 				$this->delete( $oRoom->getId() );
 				return null;
